IGNITE-15785 Move validation tests.
